avec norme init header

+++ srcs/arena/pars_header.c
@@ -20,9 +20,9 @@
 **	'prog'.
 */
 
-int		stock_file(t_champion *champ)
+int			stock_file(t_champion *champ)
 {
-	int	ret;
+	int		ret;
 	char	buf[SIZE_MAX_PROG + 1];
 	uint	min_header;
 
@@ -31,12 +31,11 @@ int		stock_file(t_champion *champ)
 	min_header = sizeof(COREWAR_EXEC_MAGIC) + PROG_NAME_LENGTH + PADDING;
 	min_header += INFO_SIZE_CODE + COMMENT_LENGTH + PADDING + 1;
 	champ->header.prog_size = ret - min_header + 1;
-// dois je mettre le +1??
 	if ((uint)ret < min_header)
-		return ((int)ft_error(TOO_SMALL, NULL));
+		return ((intptr_t)ft_error(TOO_SMALL, NULL));
 	ft_memcpy(champ->prog, buf, ret);
 	if ((ret = read(champ->fd, buf, 2)))
-		return ((int)ft_error(TOO_BIG, NULL));
+		return ((intptr_t)ft_error(TOO_BIG, NULL));
 	return (TRUE);
 }
 
@@ -83,7 +82,7 @@ int32_t		string_to_int(t_champion *champ, int size, int i)
 int			name_size_comment(t_champion *champ)
 {
 	int		i;
-	uint		nb;
+	uint	nb;
 	int		total_name;
 	int		total_comment;
 
@@ -95,7 +94,7 @@ int			name_size_comment(t_champion *champ)
 	ft_memcpy(champ->header.prog_name, champ->prog + i, total_name);
 	nb = string_to_int(champ, INFO_SIZE_CODE, i + total_name);
 	if (nb != champ->header.prog_size)
-		return ((int)ft_error(WRONG_PROG_SIZE, NULL));
+		return ((intptr_t)ft_error(WRONG_PROG_SIZE, NULL));
 	i += total_name + INFO_SIZE_CODE;
 	ft_memcpy(champ->header.comment, champ->prog + i, total_comment);
 	return (TRUE);
@@ -109,14 +108,14 @@ int			name_size_comment(t_champion *champ)
 
 int			pars_header(t_champion *champ)
 {
-	int			nb;
+	int		nb;
 
 	ft_bzero(champ->prog, SIZE_MAX_PROG);
 	if (stock_file(champ) == FALSE)
 		return (FALSE);
 	nb = string_to_int(champ, sizeof(COREWAR_EXEC_MAGIC), 0);
 	if (nb != COREWAR_EXEC_MAGIC)
-		return ((int)ft_error(WRONG_MAGIC_NB, NULL));
+		return ((intptr_t)ft_error(WRONG_MAGIC_NB, NULL));
 	else
 		champ->header.magic = nb;
 	if (name_size_comment(champ) == FALSE)
