+++ srcs/asm/cor_file.c
@@ -23,11 +23,11 @@ int			init_file(t_file *out_file, char *source_file)
 
 	l = ft_strlen(source_file);
 	if (l < 3 || ft_strcmp(source_file + l - 2, ".s"))
-		return ((int)ft_error(WRONG_SOURCE, NULL));
+		return ((intptr_t)ft_error(WRONG_SOURCE, NULL));
 	if (!(out_file->name = ft_memalloc(sizeof(char) * (l + 2))))
-		return ((int)ft_error(MALLOC_FAIL, NULL));
+		return ((intptr_t)ft_error(MALLOC_FAIL, NULL));
 	if (!(out_file->content = ft_memalloc(sizeof(char) * SIZE_MAX_PROG)))
-		return ((int)ft_error(MALLOC_FAIL, NULL));
+		return ((intptr_t)ft_error(MALLOC_FAIL, NULL));
 	out_file->name = ft_memcpy(out_file->name, source_file, l - 1);
 	out_file->name = ft_stricat(out_file->name, "cor", l - 1);
 	return (TRUE);
@@ -53,18 +53,18 @@ int			parsing_header(t_stack *stack, int fd, t_s *s)
 	s->i = 0;
 	l_name = PROG_NAME_LENGTH;
 	if (!(stack->champion_name = ft_memalloc(sizeof(char) * l_name + 1)))
-		return ((int)ft_error(MALLOC_NAME, NULL));
+		return ((intptr_t)ft_error(MALLOC_NAME, NULL));
 	if (!(stack->comment = ft_memalloc(sizeof(char) * COMMENT_LENGTH + 1)))
-		return ((int)ft_error(MALLOC_COMMENT, NULL));
+		return ((intptr_t)ft_error(MALLOC_COMMENT, NULL));
 	if (!header_content(stack, fd, s))
 	{
 		ft_memdel((void**)&s->line);
-		return ((int)just_free(stack->champion_name, stack->comment));
+		return ((intptr_t)just_free(stack->champion_name, stack->comment));
 	}
 	if (!header_content(stack, fd, s))
 	{
 		ft_memdel((void**)&s->line);
-		return ((int)just_free(stack->champion_name, stack->comment));
+		return ((intptr_t)just_free(stack->champion_name, stack->comment));
 	}
 	return (TRUE);
 }
@@ -123,7 +123,7 @@ int			fill_opcode(t_file *out_file, t_stack stack)
 		}
 		i++;
 		if (write_op_values(out_file, &i, op, stack) == FALSE)
-			return ((int)free_op_lab(&stack));
+			return ((intptr_t)free_op_lab(&stack));
 		op = op->next;
 	}
 	return (TRUE);
@@ -147,13 +147,13 @@ int			cor_file(char *source_file, t_file *out_file, int fd)
 	if (init_file(out_file, source_file) == FALSE)
 		return (FALSE);
 	if (parsing_header(&stack, fd, &s) == FALSE)
-		return ((int)just_free(out_file->name, out_file->content));
+		return ((intptr_t)just_free(out_file->name, out_file->content));
 	fill_header(out_file, &stack);
 	stack.cur_octet = out_file->total_size;
 	if (parsing_exec(&stack, fd, &s) == FALSE)
-		return ((int)just_free(out_file->name, out_file->content));
+		return ((intptr_t)just_free(out_file->name, out_file->content));
 	if (fill_opcode(out_file, stack) == FALSE)
-		return ((int)just_free(out_file->name, out_file->content));
+		return ((intptr_t)just_free(out_file->name, out_file->content));
 	real_prog_size = out_file->total_size - SIZE_HEADER;
 	nb_to_binary(out_file, INFO_PROG, out_file->prog_size, real_prog_size);
 	out_file->total_size -= INFO_PROG;
@@ -162,7 +162,7 @@ int			cor_file(char *source_file, t_file *out_file, int fd)
 	if (out_file->fd <= 0)
 	{
 		ft_error(CREATE_FAIL, out_file->name);
-		return ((int)just_free(out_file->name, out_file->content));
+		return ((intptr_t)just_free(out_file->name, out_file->content));
 	}
 	return (TRUE);
 }
