+++ srcs/asm/header_content.c
@@ -25,7 +25,7 @@ int		check_after_quote(t_s *s, t_token *token)
 		if (diff(s->line[s->i], SPACE))
 		{
 			fill_token(s, 0, token);
-			return ((int)token_free(SYNTAXE_ERROR, token));
+			return ((intptr_t)token_free(SYNTAXE_ERROR, token));
 		}
 		s->i += 1;
 	}
@@ -44,7 +44,7 @@ int		fill_empty_line(t_s *s, int fd)
 	ft_memdel((void**)&s->line);
 	ret = gnl(fd, &s->line);
 	if (ret <= 0 || s->line == NULL)
-		return ((int)token_free(MISSING_QUOTE, &token));
+		return ((intptr_t)token_free(MISSING_QUOTE, &token));
 	s->l += 1;
 	s->i = 0;
 	return (TRUE);
@@ -60,10 +60,10 @@ int		check_empty_line(t_s *s, char **tmp, int fd)
 	if (s->line[s->i] == '\0')
 	{
 		if (fill_empty_line(s, fd) == FALSE)
-			return ((int)just_free(*tmp, NULL));
+			return ((intptr_t)just_free(*tmp, NULL));
 		while (s->line != NULL && s->line[0] == '\0')
 			if (fill_empty_line(s, fd) == FALSE)
-				return ((int)just_free(*tmp, NULL));
+				return ((intptr_t)just_free(*tmp, NULL));
 	}
 	return (TRUE);
 }
@@ -134,7 +134,7 @@ int		header_content(t_stack *stack, int fd, t_s *s)
 	if (s->l == FALSE)
 		return (FALSE);
 	if (fill_name_com(type, tmp, stack, &token) == FALSE)
-		return ((int)just_free(tmp, NULL));
+		return ((intptr_t)just_free(tmp, NULL));
 	ft_memdel((void**)&tmp);
 	ft_memdel((void**)&s->line);
 	return (TRUE);
