+++ srcs/asm/header_type.c
@@ -22,13 +22,13 @@ int		fill_name_com(int type, char *tmp, t_stack *stack, t_token *token)
 	if (type == 'n')
 	{
 		if (ft_strlen(tmp) > PROG_NAME_LENGTH)
-			return ((int)token_free(TOO_LONG_NAME, token));
+			return ((intptr_t)token_free(TOO_LONG_NAME, token));
 		stack->champion_name = ft_strcpy(stack->champion_name, tmp);
 	}
 	else if (type == 'c')
 	{
 		if (ft_strlen(tmp) > COMMENT_LENGTH)
-			return ((int)token_free(TOO_LONG_COM, token));
+			return ((intptr_t)token_free(TOO_LONG_COM, token));
 		stack->comment = ft_strcpy(stack->comment, tmp);
 	}
 	return (TRUE);
@@ -46,15 +46,15 @@ int		is_valid_command_start(t_s *s, int *type, t_token *token)
 	if (*type == 'n')
 	{
 		if (ft_strncmp(s->line + s->i, NAME_CMD_STRING, 5) != 0)
-			return ((int)token_free(WRONG_HEADER, token));
+			return ((intptr_t)token_free(WRONG_HEADER, token));
 	}
 	else if (*type == 'c')
 	{
 		if (ft_strncmp(s->line + s->i, COMMENT_CMD_STRING, 8) != 0)
-			return ((int)token_free(WRONG_HEADER, token));
+			return ((intptr_t)token_free(WRONG_HEADER, token));
 	}
 	else
-		return ((int)token_free(INVALID_COMMAND, token));
+		return ((intptr_t)token_free(INVALID_COMMAND, token));
 	return (TRUE);
 }
 
@@ -70,11 +70,11 @@ int		is_valid_command_end(t_s *s, int *type, t_token *token)
 	while (s->line[s->i] != '\0' && s->line[s->i] != '"')
 	{
 		if (diff(s->line[s->i], SPACE) == TRUE)
-			return ((int)token_free(WRONG_FORMAT, token));
+			return ((intptr_t)token_free(WRONG_FORMAT, token));
 		s->i += 1;
 	}
 	if (s->line[s->i] != '"')
-		return ((int)token_free(WRONG_FORMAT, token));
+		return ((intptr_t)token_free(WRONG_FORMAT, token));
 	s->i += 1;
 	ft_memdel((void**)&token->name);
 	return (TRUE);
@@ -101,7 +101,7 @@ int		get_command_type(int fd, t_s *s, int *type, t_token *token)
 		{
 			if ((diff(s->line[s->i], SPACE) == TRUE)
 				&& fill_token(s, 0, token))
-				return ((int)token_free(WRONG_HEADER, token));
+				return ((intptr_t)token_free(WRONG_HEADER, token));
 			s->i += 1;
 		}
 		if (s->line[s->i] == '.')
@@ -110,7 +110,7 @@ int		get_command_type(int fd, t_s *s, int *type, t_token *token)
 		ft_memdel((void**)&token->name);
 	}
 	if (s->line == NULL && (token->line += 1))
-		return ((int)token_free(INCOMPLETE, token));
+		return ((intptr_t)token_free(INCOMPLETE, token));
 	*type = s->line[s->i + 1];
 	if (is_valid_command_start(s, type, token) == FALSE
 		|| is_valid_command_end(s, type, token) == FALSE)
